pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sillychess"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "torch",
  "python-chess",
  "zstandard",
  "tqdm",
]

[tool.setuptools]
packages = ["sillychess"]
